% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khiSquarePerCell.R
\name{Khi.Square.Per.Cell}
\alias{Khi.Square.Per.Cell}
\title{Khi Square Per Cell}
\usage{
Khi.Square.Per.Cell(df,
  n.Segmentation,
  decimais = 3,
  decimaisPorcentage = 0,
  dec.Global.p_Value = 6)
}
\arguments{
\item{df}{is a dataframe}

\item{n.Segmentation}{TRUE if exits in df the total number of case per each column}

\item{decimais}{the number of decimals for the results}

\item{decimaisPorcentage}{decimals used for percentage results}

\item{dec.Global.p_Value}{decimals for the global khi square reult}
}
\value{
Khi Square Per Cell in different presentation options
}
\description{
It makes Khi Square Per Cell for dataframe with the total number of case per column
}
